% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{refresh}
\alias{refresh}
\title{Refresh raw and cleaned data}
\usage{
refresh(verbose = TRUE)
}
\arguments{
\item{verbose}{Should messages about what's going on be displayed?}
}
\value{
Update the raw and cleaned country and regional CSVs of data in \code{data-raw}.
}
\description{
Refresh raw and cleaned data
}
\examples{
refresh()
}
