% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france_country}
\alias{france_country}
\title{Averages of COVID-19 cases in the country of France}
\format{
A data frame with one row per date and 8 variables:
\describe{
  \item{date}{Date}
  \item{region_name}{France}
  \item{region_code}{FRA}
  \item{confirmed}{Average number of cumulative confirmed cases on this day.}
  \item{confirmed}{Average number of cumulative dead cases on this day.}
  \item{confirmed}{Average number of cumulative hospitalized cases on this day.}
  \item{confirmed}{Average number of cumulative dead recovered on this day.}
}
}
\source{
\url{https://raw.githubusercontent.com/opencovid19-fr/data/master/dist/chiffres-cles.csv}
}
\usage{
france_country
}
\description{
One row per day summarizing the data in \code{data-raw/<date>_france_clean_country.csv}
}
\keyword{datasets}
