% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france_regional}
\alias{france_regional}
\title{Averages of COVID-19 cases per region in France}
\format{
A data frame with one row per date and 8 variables:
\describe{
  \item{date}{Date}
  \item{region_name}{Name of the region}
  \item{region_code}{Abbreviation of the}
  \item{confirmed}{Average number of cumulative confirmed cases on this day.}
  \item{confirmed}{Average number of cumulative dead cases on this day.}
  \item{confirmed}{Average number of cumulative hospitalized cases on this day.}
  \item{confirmed}{Average number of cumulative dead recovered on this day.}
}
}
\source{
\url{https://raw.githubusercontent.com/opencovid19-fr/data/master/dist/chiffres-cles.csv}
}
\usage{
france_regional
}
\description{
One row per day summarizing the data in \code{data-raw/<date>_france_clean_regional.csv}
}
\keyword{datasets}
