% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh_covid19france}
\alias{refresh_covid19france}
\title{Refresh data}
\usage{
refresh_covid19france(verbose = TRUE)
}
\arguments{
\item{verbose}{Should messages about what's going on be displayed?}
}
\value{
Get the most up to do date per-day data
}
\description{
Refresh data
}
\examples{
\donttest{
refresh_covid19france()
}
}
